Removed all sentences ending in "?" from summary text.
ALSO! I'm testing this on a portion of my psych textbook that describes
Milgram's torture experiments. Will not compile because file paths not
same. If curious here is the portion I'm testing:

"Compliance involves a change in behavior in response to a request. In
the case of obedience, the behavior change comes in response to a
demand from an authority figure (Kenrick, Neuberg, & Cialdini, 2010).
In the 1960s, Stanley Milgram developed a laboratory procedure at Yale
University to study obedience. In his first experiment, he used
newspaper ads to recruit forty male volunteers between the ages of 20
and 50. Among the participants were professionals, white-collar
businessmen, and unskilled workers (Milgram, 1963).
Imagine that you are one of the people who answered the ad. When you
arrive for the experiment, you join a 50-year-old gentleman who has
also volunteered and has been scheduled for the same session. The
experimenter explains that the purpose of the experiment is to examine
the effects of punishment on learning. One of you—the
￼￼obedience Changingbehaviorin response to a demand from an authority
figure.
￼￼© Ariel Skelley/Corbis
738
FIGURE 18.2
chapter 
Social Influence
Results of Milgram’s Obedience
Experiment 40 When Stanley Milgram asked a group 35 of undergraduates
and a group of
psychiatrists to predict how participants
in his experiment would respond, they
estimated that no more than 2 percent
would go all the way to 450 volts. In fact,
65 percent of the participants did so. What
do you think you would have done in this
situation?
Source: From S. Milgram. “Behavioral Study of Obedience.” Journal of
Abnormal and Social Psychology, 67, pp. 371–378. Reprinted by
permission.
“teacher”—will help the “learner” remember a list of words by
administering an elec- trical shock whenever he makes a mistake. Then
the experimenter turns to you and asks you to draw one of two cards out
of a hat. Your card says “Teacher.” You think to yourself that this
must be your lucky day.
Now the learner is taken into another room and strapped into a chair.
Electrodes are attached to his arms. Meanwhile, you are shown a shock
generator featuring thirty switches. The experimenter explains that the
switch on the far left administers a mild, 15-volt shock and that each
succeeding switch increases the shock by 15 volts. The switch on the
far right delivers 450 volts. The far-left section of the shock
generator is labeled “slight shock.” Looking across the panel, you see
“moderate shock,” “very strong shock,” and at the far right,
“danger—severe shock.” The last two switches are ominously labeled
“XXX.” The experimenter explains that you, the teacher, will begin by
reading a list of word pairs to the learner. Then you will go through
the list again, presenting just one word of each pair. It is the
learner’s task to say which word went with it. After the first mistake,
you are to throw the switch to deliver 15 volts of shock. Each time the
learner makes another mistake, you are to increase the shock by 15
volts.
You begin, following the experimenter’s instructions. But after the
learner makes his fifth mistake and you throw the switch to give him 75
volts, you hear a loud moan. At 90 volts, the learner cries out in
pain. At 150 volts, he screams and asks to be let out of the
experiment. You look to the experimenter, who says, “Proceed with the
next word.”
No shock was actually delivered in Milgram’s experiments. The learner
was always an employee of the experimenter, and the moans and other
sounds of pain came from a prerecorded tape. But you don’t know that.
What would you do in this situation? Suppose that you continue and
eventually deliver 180 volts. At that point, the learner screams that
he cannot stand the pain any longer and starts banging on the wall. The
experimenter says, “You have no other choice; you must go on.” Would
you continue? Would you keep going even when the learner begged to be
let out of the experiment and then fell silent? Would you administer
450 volts of potentially deadly shock to a perfect stranger just
because an experimenter demanded that you do so?
Figure 18.2 shows that only five of the forty participants in Milgram’s
experi- ment stopped before 300 volts, and twenty-six participants (65
percent) went all the way to the 450-volt level. The decision to
continue was difficult and stressful for the participants. "